add chapter 2 and chapter 3 files inside respective folders
