+++ NonblockingIO/SelectorDemo.java
@@ -1,5 +1,3 @@
-package NonblockingIO;
-
 import java.io.IOException;
 import java.net.InetSocketAddress;
 import java.nio.ByteBuffer;
