+++ URLsAndURIs/SplittingUri.java
