+++ URLsAndURIs/ToURIMethod.java
