+++ URLsAndURIs/URIParts.java
